// Frontend/library-app-web/src/models/error/ErrorType.ts
export enum ErrorType {
    REQUEST_NOT_AUTHORIZED = "REQUEST_NOT_AUTHORIZED",
    REQUEST_NOT_VALID = "REQUEST_NOT_VALID",
    RECORD_NOT_FOUND = "RECORD_NOT_FOUND",
    RECORD_NOT_CURRENT_VERSION = "RECORD_NOT_CURRENT_VERSION",
    RECORD_NOT_VALID = "RECORD_NOT_VALID",
    SYSTEM_ERROR = "SYSTEM_ERROR"
}